// control/shared/specs/db_documents/control.d.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

export interface ControlState {
  meta: {
    version: "2022.5.14";
    [k: string]: unknown;
  };
  /**
   * Generated random id representing the current session
   */
  sessionId: string;
  /**
   * Version number of control docs issued in the current session. Starts at 0
   */
  v: number;
  /**
   * Command issue epoch utc time (ms)
   */
  utcEpoch: number;
  webrtc?: {
    sdp: string;
  };
  mode?: "SAFE" | "MANUAL" | "AUTO";
}